+++ src/llmtuner/model/__init__.py
@@ -1,3 +1,5 @@
+# Level: loader > adapter > parser, utils
+
 from llmtuner.model.loader import load_model_and_tokenizer
 from llmtuner.model.parser import get_train_args, get_infer_args, get_eval_args
-from llmtuner.model.utils import create_ref_model, create_reward_model, dispatch_model, generate_model_card
+from llmtuner.model.utils import dispatch_model, generate_model_card, load_valuehead_params
